fix: move selector back to tradeQuoteSlice
